requirements.txt
# Core BigQuery & GCP libraries
google-cloud-bigquery==3.25.0
google-cloud-storage==2.19.0
db-dtypes==1.3.1
pyarrow==19.0.0

# Data processing (Python 3.13 compatible)
pandas==2.2.3
numpy==2.2.2

# Configuration
PyYAML==6.0.2
pydantic==2.10.6

# Cloud Functions runtime (used when deployed to GCP)
functions-framework==3.8.2
flask==3.1.0